make BottomOverlayBox component reusable for flight and hotel pages
